requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
google-genai
python-dotenv
pydantic